refactor: remove typings folder
