+++ src/books/LikedByUsersBook.ts
@@ -1,10 +1,10 @@
 import { Collection } from '../util';
-import { BaseBook } from './BaseBook';
+import { BaseBook, type BaseBookOptions } from './BaseBook';
 import { CustomError } from '../errors';
 import { RequestData, type User } from '../structures';
 import type { Client } from '../client';
-import type { LikedByUsersBookOptions } from '../typings';
 import type { GETTweetsIdLikingUsersQuery, GETTweetsIdLikingUsersResponse } from 'twitter-types';
+import type { TweetResolvable } from '../managers';
 
 /**
  * A class for fetching users who liked a tweet
@@ -74,4 +74,8 @@ export class LikedByUsersBook extends BaseBook {
 		}
 		return likingUsers;
 	}
+}
+
+export interface LikedByUsersBookOptions extends BaseBookOptions {
+	tweet: TweetResolvable;
 }
