+++ src/index.ts
@@ -3,5 +3,4 @@ export * from './client';
 export * from './errors';
 export * from './managers';
 export * from './structures';
-export * from './typings';
 export * from './util';
