+++ src/managers/FilteredStreamRuleManager.ts
@@ -1,15 +1,8 @@
 import { Collection } from '../util';
-import { BaseManager } from './BaseManager';
+import { type BaseFetchOptions, BaseManager } from './BaseManager';
 import { CustomError, CustomTypeError } from '../errors';
 import { RequestData, FilteredStreamRule } from '../structures';
 import type { Client } from '../client';
-import type {
-	FetchFilteredStreamRuleOptions,
-	FetchFilteredStreamRulesOptions,
-	FilteredStreamRuleData,
-	FilteredStreamRuleManagerFetchResult,
-	FilteredStreamRuleResolvable,
-} from '../typings';
 import type {
 	GETTweetsSearchStreamRulesQuery,
 	GETTweetsSearchStreamRulesResponse,
@@ -26,48 +19,46 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	}
 
 	/**
-	 * Fetches one or multiple rules that are currently active.
-	 * @param options The options for fetching rules
+	 * Fetches one or multiple rules that are currently active. Fetches all active rules if `rulesOrRules` is an empty array.
+	 * @param ruleOrRules The rule or rules to fetch or an empty array
+	 * @param options An object containing optional parameters to apply
 	 * @returns A {@link FilteredStreamRule} or a {@link Collection} of them
 	 * @example
-	 * // Fetch all active rules
-	 * const rules = await client.filteredStreamRules.fetch();
-	 *
 	 * // Fetch a single active rule
-	 * const rule = await client.filteredStreamRules.fetch({ rule: '1459555165208338435' });
+	 * const rule = await client.filteredStreamRules.fetch('1459555165208338435');
 	 *
 	 * // Fetch multiple active rules
-	 * const rules = await client.filteredStreamRules.fetch({ rules: ['1459555165208338435', '1488046998351925250'] });
+	 * const rules = await client.filteredStreamRules.fetch(['1459555165208338435', '1488046998351925250']);
+	 *
+	 * // Fetch all active rules
+	 * const rules = await client.filteredStreamRules.fetch([]);
 	 */
-	async fetch<T extends FetchFilteredStreamRuleOptions | FetchFilteredStreamRulesOptions>(
-		options?: T,
-	): Promise<FilteredStreamRuleManagerFetchResult<T>> {
-		if (typeof options === 'undefined') {
-			return this.#fetchMultipleRules() as Promise<FilteredStreamRuleManagerFetchResult<T>>;
-		} else if ('rule' in options) {
-			const ruleId = this.resolveId(options.rule);
-			if (!ruleId) throw new CustomError('RULE_RESOLVE_ID', 'fetch');
-			return this.#fetchSingleRule(ruleId, options) as Promise<FilteredStreamRuleManagerFetchResult<T>>;
-		} else if ('rules' in options) {
-			if (typeof options.rules === 'undefined') {
-				return this.#fetchMultipleRules(options.rules, options) as Promise<FilteredStreamRuleManagerFetchResult<T>>;
+	async fetch<R extends FilteredStreamRuleResolvable | Array<FilteredStreamRuleResolvable>>(
+		ruleOrRules: R,
+		options?: FetchFilteredStreamRuleOrRulesOptions<R>,
+	): Promise<FilteredStreamRuleManagerFetchResult<R>> {
+		if (Array.isArray(ruleOrRules)) {
+			if (ruleOrRules.length) {
+				const ruleIds = ruleOrRules.map(rule => {
+					const ruleId = this.resolveId(rule);
+					if (!ruleId) throw new CustomError('RULE_RESOLVE_ID', 'fetch');
+					return ruleId;
+				});
+				// @ts-expect-error TS seems to not work when conditional types and promises are combined together
+				return this.#fetchMultipleRulesByIds(ruleIds, options);
 			}
-			if (!Array.isArray(options.rules)) {
-				throw new CustomTypeError('INVALID_TYPE', 'rules', 'array or undefined', true);
-			}
-			const ruleIds = options.rules.map(rule => {
-				const ruleId = this.resolveId(rule);
-				if (!ruleId) throw new CustomError('RULE_RESOLVE_ID', 'fetch');
-				return ruleId;
-			});
-			return this.#fetchMultipleRules(ruleIds, options) as Promise<FilteredStreamRuleManagerFetchResult<T>>;
+			// @ts-expect-error TS seems to not work when conditional types and promises are combined together
+			return this.#fetchMultipleRulesByIds([], options);
 		}
-		return this.#fetchMultipleRules() as Promise<FilteredStreamRuleManagerFetchResult<T>>;
+		const ruleId = this.resolveId(ruleOrRules);
+		if (!ruleId) throw new CustomError('RULE_RESOLVE_ID', 'fetch');
+		// @ts-expect-error TS seems to not work when conditional types and promises are combined together
+		return this.#fetchSingleRuleById(ruleId, options);
 	}
 
 	/**
-	 * Creates rules for the filtered stream.
-	 * @param data The data for creating rules
+	 * Creates one or more rules for the filtered stream.
+	 * @param data The data for creating one or more rules
 	 * @returns A {@link Collection} of {@link FilteredStreamRule}
 	 * @example
 	 * // Create a single rule
@@ -99,8 +90,8 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	}
 
 	/**
-	 * Deletes rules for the filtered stream using their ids.
-	 * @param ruleId The id or ids of the rules to delete
+	 * Deletes one or more filtered stream rules using their ids.
+	 * @param ruleIdOrIds The id or ids of the rules to delete
 	 * @example
 	 * // Delete a single rule
 	 * const data = await client.filteredStreamRules.deleteById('1488053806785245187');
@@ -108,8 +99,8 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	 * // Delete multiple rules
 	 * const data = await client.filteredStreamRules.deleteById(['1488048453506957314', '1488053806785245186']);
 	 */
-	async deleteById(ruleId: string | Array<string>): Promise<POSTTweetsSearchStreamRulesResponse> {
-		const ids = Array.isArray(ruleId) ? ruleId : [ruleId];
+	async deleteById(ruleIdOrIds: string | Array<string>): Promise<POSTTweetsSearchStreamRulesResponse> {
+		const ids = Array.isArray(ruleIdOrIds) ? ruleIdOrIds : [ruleIdOrIds];
 		const body: POSTTweetsSearchStreamRulesJSONBody = {
 			delete: {
 				ids,
@@ -119,8 +110,8 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	}
 
 	/**
-	 * Deletes rules for the filtered stream using their values.
-	 * @param ruleValue The value or values of the rules to delete
+	 * Deletes one or more filtered stream rules using their values.
+	 * @param ruleValueOrValues The value or values of the rules to delete
 	 * @example
 	 * // Delete a single rule
 	 * const data = await client.filteredStreamRules.deleteByValue('@iShiibi');
@@ -128,8 +119,8 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	 * // Delete multiple rules
 	 * const data = await client.filteredStreamRules.deleteByValue(['from:iShiibi', 'to:TwitterAPI']);
 	 */
-	async deleteByValue(ruleValue: string | Array<string>): Promise<POSTTweetsSearchStreamRulesResponse> {
-		const values = Array.isArray(ruleValue) ? ruleValue : [ruleValue];
+	async deleteByValue(ruleValueOrValues: string | Array<string>): Promise<POSTTweetsSearchStreamRulesResponse> {
+		const values = Array.isArray(ruleValueOrValues) ? ruleValueOrValues : [ruleValueOrValues];
 		const body: POSTTweetsSearchStreamRulesJSONBody = {
 			delete: {
 				values,
@@ -139,7 +130,7 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	}
 
 	/**
-	 * Delete rules for the filtered stream.
+	 * Deletes one or more filtered stream rules.
 	 * @param body The request body
 	 * @returns // TODO
 	 */
@@ -154,11 +145,11 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 	/**
 	 * Fetches a single rule by using its id.
 	 * @param ruleId The id of the rule to fetch
-	 * @param options The options for fetching the rule
+	 * @param options An object containing optional parameters to apply
 	 * @returns A {@link FilteredStreamRule}
 	 */
-	async #fetchSingleRule(ruleId: string, options: FetchFilteredStreamRuleOptions): Promise<FilteredStreamRule> {
-		if (!options.skipCacheCheck) {
+	async #fetchSingleRuleById(ruleId: string, options?: FetchFilteredStreamRuleOptions): Promise<FilteredStreamRule> {
+		if (!options?.skipCacheCheck) {
 			const cachedRule = this.cache.get(ruleId);
 			if (cachedRule) return cachedRule;
 		}
@@ -169,22 +160,23 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 		const res: GETTweetsSearchStreamRulesResponse = await this.client._api.tweets.search.stream.rules.get(requestData);
 		const rawRule = res.data?.[0];
 		if (!rawRule) throw new CustomError('RULE_NOT_FOUND');
-		return this._add(rawRule.id, rawRule, options.cacheAfterFetching);
+		return this._add(rawRule.id, rawRule, options?.cacheAfterFetching);
 	}
 
 	/**
-	 * Fetches multiple rules by using their ids.
-	 * @param ruleIds The ids of the rules to fetch
-	 * @param options The options for fetching the rules
+	 * Fetches multiple rules by using their ids or all rules if `ruleIds` is an empty array.
+	 * @param ruleIds The ids of the rules to fetch or an empty array
+	 * @param options An object containing optional parameters to apply
 	 * @returns A {@link Collection} of {@link FilteredStreamRule}
 	 */
-	async #fetchMultipleRules(
-		ruleIds?: Array<string>,
+	async #fetchMultipleRulesByIds(
+		ruleIds: Array<string>,
 		options?: FetchFilteredStreamRulesOptions,
 	): Promise<Collection<string, FilteredStreamRule>> {
+		if (!Array.isArray(ruleIds)) throw new CustomTypeError('INVALID_TYPE', 'rulesIds', 'array', true);
 		const fetchedRules = new Collection<string, FilteredStreamRule>();
 		const query: GETTweetsSearchStreamRulesQuery = {
-			ids: ruleIds,
+			ids: ruleIds.length ? ruleIds : undefined,
 		};
 		const requestData = new RequestData({ query });
 		const res: GETTweetsSearchStreamRulesResponse = await this.client._api.tweets.search.stream.rules.get(requestData);
@@ -197,3 +189,42 @@ export class FilteredStreamRuleManager extends BaseManager<string, FilteredStrea
 		return fetchedRules;
 	}
 }
+
+/**
+ * Options used to feth a single filtered stream rule
+ */
+export type FetchFilteredStreamRuleOptions = BaseFetchOptions;
+
+/**
+ * Options used to feth multiple filtered stream rules
+ */
+export type FetchFilteredStreamRulesOptions = Omit<BaseFetchOptions, 'skipCacheCheck'>;
+
+/**
+ * Options used to fetch one or more filter stream rules
+ */
+export type FetchFilteredStreamRuleOrRulesOptions<
+	R extends FilteredStreamRuleResolvable | Array<FilteredStreamRuleResolvable>,
+> = R extends FilteredStreamRuleResolvable ? FetchFilteredStreamRuleOptions : FetchFilteredStreamRulesOptions;
+
+export type FilteredStreamRuleManagerFetchResult<
+	R extends FilteredStreamRuleResolvable | Array<FilteredStreamRuleResolvable>,
+> = R extends FilteredStreamRuleResolvable ? FilteredStreamRule : Collection<string, FilteredStreamRule>;
+
+/**
+ * Options used to create a new filtered stream rule
+ */
+export interface FilteredStreamRuleData {
+	/**
+	 * The value of the rule
+	 * @see https://developer.twitter.com/en/docs/twitter-api/tweets/search/integrate/build-a-query
+	 */
+	value: string;
+
+	/**
+	 * The label of the rule
+	 */
+	tag?: string;
+}
+
+export type FilteredStreamRuleResolvable = FilteredStreamRule | string;
