--- src/typings/index.ts
@@ -1,2 +1,0 @@
-export * from './Interfaces';
-export * from './Types';
